@@ -64,38 +64,38 @@ class RestrictionsController < ApplicationController
   end
 
   protected
-
+  # Can't do this until we decide what to do with restrictions
   def set_sources
-    if params[:for_users]
-      params[:for_users].split(",").each do |l|
-        if l == l.split("||").first #This is for if javascript is disabled
-          l = l.strip
-          user_source = User.search(l) || Role.find_by_name(l)
-          find_dept = Department.find_by_name(l)
-          user_source = find_dept if find_dept && current_user.is_admin_of?(find_dept)
-          @restriction.user_sources << user_source if user_source
-        else
-          l = l.split("||")
-          @restriction.user_sources << l[0].constantize.find(l[1]) if l.length == 2 #javascript or not javascript
-        end
-      end
-    end
-    if params[:department_wide_locations] && current_user.is_admin_of?(current_department)
-      @restriction.location_sources << current_department
-      @restriction.loc_groups << current_department.loc_groups
-      @restriction.locations << current_department.locations
-    elsif params[:for_location_groups]
-      params[:for_location_groups].each do |loc_group|
-        lg = LocGroup.find(loc_group)
-        @restriction.loc_groups << lg
-        @restriction.locations << lg.locations
-      end
-    end
-    if params[:for_locations]
-      params[:for_locations].each do |loc|
-        @restriction.locations << Location.find_by_id(loc)
-      end
-    end
+  #   if params[:for_users]
+  #     params[:for_users].split(",").each do |l|
+  #       if l == l.split("||").first #This is for if javascript is disabled
+  #         l = l.strip
+  #         user_source = User.search(l) || Role.find_by_name(l)
+  #         find_dept = Department.find_by_name(l)
+  #         user_source = find_dept if find_dept && current_user.is_admin_of?(find_dept)
+  #         @restriction.user_sources << user_source if user_source
+  #       else
+  #         l = l.split("||")
+  #         @restriction.user_sources << l[0].constantize.find(l[1]) if l.length == 2 #javascript or not javascript
+  #       end
+  #     end
+  #   end
+  #   if params[:department_wide_locations] && current_user.is_admin_of?(current_department)
+  #     # Do we allow department wide actions? -Hugh
+  #     # @restriction.location_sources << current_department
+  #     @restriction.loc_groups << current_department.loc_groups
+  #     @restriction.locations << current_department.locations
+  #   elsif params[:for_location_groups]
+  #     params[:for_location_groups].each do |loc_group|
+  #       lg = LocGroup.find(loc_group)
+  #       @restriction.loc_groups << lg
+  #       @restriction.locations << lg.locations
+  #     end
+  #   end
+  #   if params[:for_locations]
+  #     params[:for_locations].each do |loc|
+  #       @restriction.locations << Location.find_by_id(loc)
+  #     end
+  #   end
   end
-
 end
