@@ -1,22 +1,23 @@
 module NoticesHelper
 
-  def department_checked?(dept, notice)
-   notice.departments.each do |d|
-      return true if d == dept
-    end
-    false
-  end
+  # This don't work anymore, fixing with dept-wide column in notices soon -Hugh
+  # def department_checked?(dept, notice)
+  #  notice.departments.each do |d|
+  #     return true if d == dept
+  #   end
+  #   false
+  # end
 
-  def loc_group_checked?(loc_group, notice)
+  def loc_group_checked_notice?(loc_group, notice)
     notice.loc_groups.each do |lg|
       return true if lg == loc_group
     end
     false
   end
 
-  def location_checked?(location, notice)
+  def location_checked_notice?(location, notice)
 		return @current_shift_location == location if @current_shift_location
-    notice.location_sources.each do |loc|
+    notice.display_locations.each do |loc|
       return true if loc == location
     end
 		false
