@@ -13,11 +13,17 @@ class LocGroup < ActiveRecord::Base
               :foreign_key => "admin_perm_id",
               :dependent => :destroy
   has_many :locations, :dependent => :destroy
+
+  # These make loc groups connect to the superclass notice, as well as its subclasses
+  has_and_belongs_to_many :announcements, :join_table => :loc_groups_notices, :association_foreign_key => :notice_id
+  has_and_belongs_to_many :links,         :join_table => :loc_groups_notices, :association_foreign_key => :notice_id
+  has_and_belongs_to_many :stickies,      :join_table => :loc_groups_notices, :association_foreign_key => :notice_id
+  has_and_belongs_to_many :notices
   
-  named_scope :active, :conditions => {:active => true}
+  scope :active, :conditions => {:active => true}
   
-  before_validation_on_create :create_permissions
-  before_validation_on_update :update_permissions
+  before_validation(:on => :create) {:create_permissions}
+  before_validation(:on => :update) {:update_permissions}
 
   validates_presence_of :department
 
@@ -35,7 +41,7 @@ class LocGroup < ActiveRecord::Base
   end
   
   def roles
-      department.roles.select { |u| u.permissions.include?(signup_permission) }
+    department.roles.select { |u| u.permissions.include?(signup_permission) }
   end
   
   def deactivate
