@@ -26,10 +26,10 @@ class PayformItem < ActiveRecord::Base
   validates_numericality_of :hours, :greater_than => 0
   validates_presence_of :reason, :on => :update
   validate :length_of_description
-  validate_on_update :length_of_reason
+  validate :length_of_reason, :on => :update
 
-  named_scope :active, :conditions => {:active =>  true}
-  named_scope :in_category, lambda { |category| { :conditions => {:category_id => category.id}}}
+  scope :active, :conditions => {:active =>  true}
+  scope :in_category, lambda { |category| { :conditions => {:category_id => category.id}}}
   
   def add_errors(e)
     e = e.to_s.gsub("Validation failed: ", "")
